+++ modules/tinymce/src/plugins/table/main/ts/api/Commands.ts
@@ -25,7 +25,10 @@ import * as RowDialog from '../ui/RowDialog';
 import * as TableDialog from '../ui/TableDialog';
 import { isPercentagesForced, isPixelsForced, isResponsiveForced } from './Settings';
 
+type ExecuteAction<T> = (table: SugarElement<HTMLTableElement>, startCell: SugarElement<HTMLTableCellElement>) => T;
+
 const getSelectionStartCellOrCaption = (editor: Editor) => TableSelection.getSelectionStartCellOrCaption(Util.getSelectionStart(editor), Util.getIsRoot(editor));
+
 const getSelectionStartCell = (editor: Editor) => TableSelection.getSelectionStartCell(Util.getSelectionStart(editor), Util.getIsRoot(editor));
 
 const registerCommands = (editor: Editor, actions: TableActions, cellSelection: CellSelectionApi, selections: Selections, clipboard: Clipboard) => {
@@ -69,6 +72,25 @@ const registerCommands = (editor: Editor, actions: TableActions, cellSelection: 
 
   const getTableFromCell = (cell: SugarElement<HTMLTableCellElement>) => TableLookup.table(cell, isRoot);
 
+  const performActionOnSelection = <T>(action: ExecuteAction<T>): Optional<T> => getSelectionStartCell(editor).bind((cell) => getTableFromCell(cell).map((table) => action(table, cell)));
+
+  const toggleTableClass = (_ui: boolean, clazz: string) => {
+    performActionOnSelection((table) => {
+      editor.formatter.toggle('tableclass', { value: clazz }, table.dom);
+      Events.fireTableModified(editor, table.dom, Events.styleModified);
+    });
+  };
+
+  const toggleTableCellClass = (_ui: boolean, clazz: string) => {
+    performActionOnSelection((table, startCell) => {
+      const cells = TableSelection.getCellsFromSelection(startCell, selections, isRoot);
+      Arr.each(cells, (value) => {
+        editor.formatter.toggle('tablecellclass', { value: clazz }, value.dom);
+      });
+      Events.fireTableModified(editor, table.dom, Events.styleModified);
+    });
+  };
+
   const postExecute = (table: SugarElement<HTMLTableElement>) => (data: TableActionResult): void => {
     editor.selection.setRng(data.rng);
     editor.focus();
@@ -77,38 +99,34 @@ const registerCommands = (editor: Editor, actions: TableActions, cellSelection: 
     Events.fireTableModified(editor, table.dom, data.effect);
   };
 
-  const actOnSelection = (execute: CombinedTargetsTableAction): void =>
-    getSelectionStartCell(editor).each((cell) => {
-      getTableFromCell(cell).each((table) => {
-        const targets = TableTargets.forMenu(selections, table, cell);
-        execute(table, targets).each(postExecute(table));
-      });
+  const actOnSelection = (execute: CombinedTargetsTableAction) =>
+    performActionOnSelection((table, startCell) => {
+      const targets = TableTargets.forMenu(selections, table, startCell);
+      execute(table, targets).each(postExecute(table));
     });
 
-  const copyRowSelection = () => getSelectionStartCell(editor).map((cell) =>
-    getTableFromCell(cell).bind((table) => {
-      const targets = TableTargets.forMenu(selections, table, cell);
+  const copyRowSelection = () =>
+    performActionOnSelection((table, startCell) => {
+      const targets = TableTargets.forMenu(selections, table, startCell);
       const generators = TableFill.cellOperations(Fun.noop, SugarElement.fromDom(editor.getDoc()), Optional.none());
       return CopyRows.copyRows(table, targets, generators);
-    }));
+    });
 
-  const copyColSelection = () => getSelectionStartCell(editor).map((cell) =>
-    getTableFromCell(cell).bind((table) => {
-      const targets = TableTargets.forMenu(selections, table, cell);
+  const copyColSelection = () =>
+    performActionOnSelection((table, startCell) => {
+      const targets = TableTargets.forMenu(selections, table, startCell);
       return CopyCols.copyCols(table, targets);
-    }));
+    });
 
   const pasteOnSelection = (execute: AdvancedPasteTableAction, getRows: () => Optional<SugarElement<HTMLTableRowElement | HTMLTableColElement>[]>) =>
     // If we have clipboard rows to paste
     getRows().each((rows) => {
       const clonedRows = Arr.map(rows, (row) => Replication.deep<HTMLTableColElement | HTMLTableRowElement>(row));
-      getSelectionStartCell(editor).each((cell) =>
-        getTableFromCell(cell).each((table) => {
-          const generators = TableFill.paste(SugarElement.fromDom(editor.getDoc()));
-          const targets = TableTargets.pasteRows(selections, cell, clonedRows, generators);
-          execute(table, targets).each(postExecute(table));
-        })
-      );
+      performActionOnSelection((table, startCell) => {
+        const generators = TableFill.paste(SugarElement.fromDom(editor.getDoc()));
+        const targets = TableTargets.pasteRows(selections, startCell, clonedRows, generators);
+        execute(table, targets).each(postExecute(table));
+      });
     });
 
   // Register action commands
@@ -121,22 +139,24 @@ const registerCommands = (editor: Editor, actions: TableActions, cellSelection: 
     mceTableInsertColAfter: () => actOnSelection(actions.insertColumnsAfter),
     mceTableDeleteCol: () => actOnSelection(actions.deleteColumn),
     mceTableDeleteRow: () => actOnSelection(actions.deleteRow),
-    mceTableCutCol: (_grid) => copyColSelection().each((selection) => {
+    mceTableCutCol: () => copyColSelection().each((selection) => {
       clipboard.setColumns(selection);
       actOnSelection(actions.deleteColumn);
     }),
-    mceTableCutRow: (_grid) => copyRowSelection().each((selection) => {
+    mceTableCutRow: () => copyRowSelection().each((selection) => {
       clipboard.setRows(selection);
       actOnSelection(actions.deleteRow);
     }),
-    mceTableCopyCol: (_grid) => copyColSelection().each((selection) => clipboard.setColumns(selection)),
-    mceTableCopyRow: (_grid) => copyRowSelection().each((selection) => clipboard.setRows(selection)),
-    mceTablePasteColBefore: (_grid) => pasteOnSelection(actions.pasteColsBefore, clipboard.getColumns),
-    mceTablePasteColAfter: (_grid) => pasteOnSelection(actions.pasteColsAfter, clipboard.getColumns),
-    mceTablePasteRowBefore: (_grid) => pasteOnSelection(actions.pasteRowsBefore, clipboard.getRows),
-    mceTablePasteRowAfter: (_grid) => pasteOnSelection(actions.pasteRowsAfter, clipboard.getRows),
+    mceTableCopyCol: () => copyColSelection().each((selection) => clipboard.setColumns(selection)),
+    mceTableCopyRow: () => copyRowSelection().each((selection) => clipboard.setRows(selection)),
+    mceTablePasteColBefore: () => pasteOnSelection(actions.pasteColsBefore, clipboard.getColumns),
+    mceTablePasteColAfter: () => pasteOnSelection(actions.pasteColsAfter, clipboard.getColumns),
+    mceTablePasteRowBefore: () => pasteOnSelection(actions.pasteRowsBefore, clipboard.getRows),
+    mceTablePasteRowAfter: () => pasteOnSelection(actions.pasteRowsAfter, clipboard.getRows),
     mceTableDelete: eraseTable,
-    mceTableSizingMode: (ui: boolean, sizing: string) => setSizingMode(sizing)
+    mceTableCellToggleClass: toggleTableCellClass,
+    mceTableToggleClass: toggleTableClass,
+    mceTableSizingMode: (_ui: boolean, sizing: string) => setSizingMode(sizing)
   }, (func, name) => editor.addCommand(name, func));
 
   // Due to a bug, we need to pass through a reference to the table obtained before the modification